Reorganize instructions to prioritize what we actually grade on
Since the minimum bar is the associations, I believe in fairness we should put those first in the readme so we're transparent about what students should focus on first to do their best on the assessment. I'm not attached to calling out the Algorithmic methods as stretch, but I include it for comment (I think we should unless there's a good reason not to).